requirements.txt
annotated-types==0.7.0
autopep8==2.1.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
